Removed second Nav Graph and Second Viewmodel. Reduced memory and fixed several bugs.
