+++ app/src/main/java/com/cbdn/reports/ui/views/newreport/DetailSections.kt
@@ -1,0 +1,8 @@
+package com.cbdn.reports.ui.views.newreport
+
+enum class DetailSections {
+    DispatchDetails,
+    LocationDetails,
+    SiteDetails,
+    SubmittalDetails
+}
